Move supervisor and tests into services folder
