+++ Cargo.toml
@@ -6,12 +6,12 @@ members = [
     "examples/cryptocurrency-advanced/backend",
     "examples/timestamping/backend",
     "services/time",
+    "services/supervisor",
     "testkit",
     "testkit/server",
     "components/build",
     "components/crypto",
     "components/derive",
     "components/merkledb",
-    "components/supervisor",
 ]
 exclude = [ "exonum/fuzz" ]
