+++ exonum/src/blockchain/tests.rs
@@ -193,7 +193,7 @@ trait ServiceGood {}
     proto_sources = "crate::proto::schema",
     implements("ServiceGood")
 )]
-struct ServiceGoodImpl;
+pub struct ServiceGoodImpl;
 
 impl ServiceGood for ServiceGoodImpl {}
