+++ services/supervisor/Cargo.toml
@@ -25,8 +25,11 @@ log = "0.4.6"
 hex = "0.4"
 
 exonum = { version = "0.12.0", path = "../../exonum" }
-exonum-derive = { version = "0.12.0", path = "../derive" }
-exonum-merkledb = { version = "0.12.0", path = "../merkledb" }
+exonum-derive = { version = "0.12.0", path = "../../components/derive" }
+exonum-merkledb = { version = "0.12.0", path = "../../components/merkledb" }
+
+[dev-dependencies]
+exonum-testkit = { version = "0.12.0", path = "../../testkit" }
 
 [build-dependencies]
-exonum-build = { version = "0.12.0", path = "../build" }
+exonum-build = { version = "0.12.0", path = "../../components/build" }
