+++ services/supervisor/README.md
@@ -1,0 +1,14 @@
+# exonum-supervisor
+
+[![Travis Build Status](https://img.shields.io/travis/exonum/exonum/master.svg?label=Linux%20Build)](https://travis-ci.com/exonum/exonum)
+[![License: Apache-2.0](https://img.shields.io/github/license/exonum/exonum.svg)](https://github.com/exonum/exonum/blob/master/LICENSE)
+![rust 1.33.0+ required](https://img.shields.io/badge/rust-1.33.0+-blue.svg?label=Required%20Rust)
+
+Exonum-supverisor is a main service of the [Exonum blockchain framework](https://exonum.com/).
+
+It is capable of deploying and starting new services.
+
+## License
+
+`exonum-supervisor` is licensed under the Apache License (Version 2.0).
+See [LICENSE](LICENSE) for details.
