+++ services/supervisor/src/errors.rs
