+++ testkit/Cargo.toml
@@ -32,6 +32,7 @@ serde_urlencoded = "0.6.0"
 tokio-core = "0.1.17"
 protobuf = { version = "2.8.0", features = ["with-serde"] }
 exonum-merkledb = { version = "0.12.0", path = "../components/merkledb" }
+exonum-supervisor  = { version = "0.12.0", path = "../services/supervisor" }
 
 [dev-dependencies]
 hex = "0.4"
