+++ testkit/src/lib.rs
@@ -180,6 +180,7 @@ use exonum::{
     runtime::{rust::ServiceFactory, InstanceId},
 };
 use exonum_merkledb::{Database, ObjectHash, Patch, Snapshot, TemporaryDB};
+use exonum_supervisor::Supervisor;
 use futures::{sync::mpsc, Future, Stream};
 use tokio_core::reactor::Core;
 
@@ -242,6 +243,19 @@ impl TestKit {
             .create()
     }
 
+    /// Returns the list of the Instances enabled by default.
+    ///
+    /// Currently only `Supervisor` service is included.
+    fn default_service_instances() -> Vec<InstanceCollection> {
+        let supervisor_instance = InstanceCollection::new(Supervisor).with_instance(
+            Supervisor::BUILTIN_ID,
+            Supervisor::BUILTIN_NAME,
+            (),
+        );
+
+        vec![supervisor_instance]
+    }
+
     fn assemble(
         database: impl Into<CheckpointDb<TemporaryDB>>,
         service_factories: impl IntoIterator<Item = InstanceCollection>,
@@ -251,12 +265,15 @@ impl TestKit {
         let api_channel = mpsc::channel(1_000);
         let api_sender = ApiSender::new(api_channel.0.clone());
 
+        let default_instances = TestKit::default_service_instances();
+        let overall_instances = default_instances.into_iter().chain(service_factories);
+
         let db = database.into();
 
         let db_handler = db.handler();
         let blockchain = Blockchain::new(
             db,
-            service_factories,
+            overall_instances,
             genesis,
             network.us().service_keypair(),
             api_sender.clone(),
